requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.32.1
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
psutil==5.9.6
httpx==0.25.2
